Add files via upload
data visualization using looker studio with five visualizations namely, revenue by store location, product category  and revenue,  revenue by month, total revenue by time buckets and month_id, with June generating the most revenue of above 165000 and even in with the time bucket the most was sold in June and in the evening where the sales are low. Thursday is the lowest selling days of the week generating less  in all categories. Saturdays  brings the best sales  of the week all though out the day.  The Hell's kitchen is the top selling  store among the  three store